requirements.txt
flask
flask-cors
google-genai
pillow
numpy
gunicorn
flask-sqlalchemy
psycopg2-binary
